Don't unmanage windows that are unmapped
Some clients keep their windows around after unmapping them (e.g. applications
that have systray icons) so that they can bring up the window without creating
a new one and doing a new MapRequest. If we unmanage() these, we'll have
forgotten about the window when we see the ConfigureRequest and not manage it
correctly.

This is a hunch about #477