Check the results of ConfigureWindow requests
This fixes a similar issue that 71a8d0d04602c53dd05e15a3cf8dde0a00e1fc97 does.
In some cases, we have 'dead' windows that we attempt to operate on. If we just
send unchecked requests, we get the exceptions back in the poll loop and
they're rendered in the users' logs. Instead, we should check them and fail
immediately. The previous patch changes the logging scheme slightly so that
these kinds of errors are not rendered.

Closes #539