add the notion of finalizers
In particular, avoid the use of __del__ as a finalizer. This has several
problems: it isn't necessarily run at all, and if it is, it isn't necessarily
run at a specific point in time (i.e. when the xcb connection is still open).

I belive this should also fix the segfault, and prevent us from leaking the
pixmaps drawer allocates when we exec.

Possible fix for #658