pangocffi: get rid of our CairoContext
we don't really have a local CairoContext. I think I just put this in
because I was on a rampage trying to refactor things to use cffi. Instead,
just use the cairo context as-is, and avoid adding this silly wrapper. this
will allow us to mess with the actual cairo context in the future more
easily.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>