group: add skip_managed parameter to toscreen()
People often ask if there's a way to handle multi monitor setups better,
finding it confusing that qtile will switch to a group that is already
managed on a different screen.

Roger points out that we have a skip_managed parameter for next_group(),
but most configs use toscreen() as well to call particular groups to a
screen. So let's add a similar parameter to that.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>