xcore: don't require XCore object for get_{keys/modifiers}
get_{keys,modifiers} currently return a hard coded list of keysyms and
modifiers, so there's no reason to require a connection to the X server.
(This could change in the future if someone implements support for querying
the server for keysyms, but nobody has yet :)

We'll use the validate() method in the `qtile check` command, and requiring
a connection to an X server seems very heavy weight (e.g. when fixing and
validating a config from a tty because it was previously broken, having a
dependency on an X server being present would be annoying).

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>