x11: prefer randr instead of xinerama for screen geometries
xinerama is quite old, and most people use randr these days to set their
screen geometries. if we use xinerama to query screen geometries, that will
not be accurately reflected if users have used randr.

unfortunately, we cannot actually delete xinerama, since Xephyr does not
seem to really support randr (or at least, I played around with it with
-xinerama +extensions RANDR and couldn't get it to show me two screens in
the tests), so let's switch the order of preference to use randr first if
it is available, and explicitly disable randr in our test suite.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>