try to type check configurable args
inspired by this reddit thread (qtile has a reddit?):

https://www.reddit.com/r/qtile/comments/1kxd5dt/trying_out_qtile_some_problems/

I noted that `qtile check` said this person's config was okay, but in fact
it is not: they had mixed up some types, which we could have detected, but
we don't have typed defaults.

Maybe we should go through and type our defaults, but for now, this would
have caught the above confusion and maybe a few others.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>